backend/pyproject.toml
[project]
name = "equipment-tracker-backend"
version = "0.1.0"
description = "Hospital Equipment Tracker Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "langgraph>=0.5.3",
    "openai>=1.96.0",
    "langchain-openai>=0.3.28",
    "python-dotenv>=1.1.1",
]
